写完所有联立查询接口和分页，待测试和待修改已测得的bug

+++ umeijia/src/main/java/com/umeijia/dao/StudentDao.java
@@ -89,7 +89,7 @@ public class StudentDao {
         }
     }
 
-    public List<Student> queryStudentByClassIdAndStudentName(int classId,String name){
+    public List<Student> queryStudentByClassIdAndStudentName(long classId,String name){
         Session session = DBManager.getSession();
         session.clear();
         String sql = String.format("from Student as u where u.cla.id=%d and u.name=%s and u.valid=1", classId,name);
